최종 버전

See merge request s04-webmobile3-sub3/s04p13a106!8

+++ backend/src/main/java/com/ssafy/project/dto/MainUnit.java
@@ -19,7 +19,7 @@ public class MainUnit {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer mainUnitId;
+	private Long mainUnitId;
 	
 	@ManyToOne
 	@JoinColumn(name = "book")
