+++ backend/src/main/java/com/ssafy/project/repository/SubUnitRepository.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.project.dto.SubUnit;
 
 @Repository
-public interface SubUnitRepository extends JpaRepository<SubUnit, Integer> {
+public interface SubUnitRepository extends JpaRepository<SubUnit, Long> {
 
 	@Query(value="select * from subunit where primarys=:id", nativeQuery=true)
-	List<SubUnit> findByUnitOf(@Param("id") Integer id);
+	List<SubUnit> findByUnitOf(@Param("id") Long id);
 }
